feat(desktop): simplify release process to tag-and-push
Move version bumping into CI so the release flow is just
`git tag desktop/v<version> && git push`. CI extracts the version from
the tag, patches package.json / tauri.conf.json / Cargo.toml, regenerates
Cargo.lock, then builds, signs, and publishes.

- Add set-version-from-tag.mjs (pure file-patcher, replaces bump-version.mjs)
- Add RELEASE_VERSION job-level env var as single source of truth in CI
- Add cargo generate-lockfile step after version patching
- Simplify justfile: kill desktop-prepare, desktop-release is now tag+push
- Add desktop-set-version and optional version arg to desktop-release-build
- Rewrite RELEASING.md for the simplified flow
